requirements.txt
psycopg2 
faker
streamlit
langchain
sqlalchemy
psycopg2-binary
google-generativeai
pinecone
tiktoken

langchain-community
google-generativeai

python-dotenv
google-generativeai
langchain-community
langchain-google-genai
psycopg2-binary
python-dotenv
rich
sqlalchemy
streamlit
